change: rename integration test package from main to integration
