+++ integration/getting_started_single_process_config_test.go
@@ -1,6 +1,6 @@
 // +build requires_docker
 
-package main
+package integration
 
 import (
 	"fmt"
