+++ integration/integration_memberlist_single_binary_test.go
@@ -1,6 +1,6 @@
 // +build requires_docker
 
-package main
+package integration
 
 import (
 	"testing"
