+++ integration/querier_streaming_mixed_ingester_test.go
@@ -1,6 +1,6 @@
 // +build requires_docker
 
-package main
+package integration
 
 import (
 	"context"
